moved cmd_tasks.c to main directory - had weird problem w/SEND_CLIENT_LIST

+++ thread_client/do1.sh
@@ -1,5 +1,7 @@
-
-make -f ts4600_$1.mak clean
+clear
+#make -f ts4600_$1.mak clean
+rm *.o &> out.txt
+rm sched150 &> out.txt
 make -f ts4600_$1.mak &> out.txt
 if grep -q error out.txt
  then
@@ -12,6 +14,12 @@ if grep -q undefined out.txt
   find2 undefined out.txt
   exit 1
 fi
-rm *.o
-mv sched$1 /home/dan/dev/sched/sched
-ls -ltr /home/dan/dev/sched
+if [ -e sched$1 ]
+ then
+   rm *.o
+   mv sched$1 /home/dan/dev/sched/sched
+   ls -ltr /home/dan/dev/sched
+   exit 1
+fi
+echo "something went wrong..."
+cat out.txt
